refactor: split stack definitions into one file per stack
