+++ pkg/stack/pgadmin.go
@@ -1,0 +1,30 @@
+package stack
+
+import "github.com/dhis2-sre/im-manager/pkg/kube"
+
+// Stack representing ../../stacks/pgadmin/helmfile.yaml.gotmpl
+var PgAdmin = Stack{
+	Name: "pgadmin",
+	Parameters: StackParameters{
+		"PGADMIN_USERNAME":  {Priority: 1, DisplayName: "pgAdmin Username", Sensitive: true},
+		"PGADMIN_PASSWORD":  {Priority: 2, DisplayName: "pgAdmin Password", Sensitive: true},
+		"CHART_VERSION":     {Priority: 3, DisplayName: "Chart Version", DefaultValue: &pgAdminDefaults.chartVersion},
+		"DATABASE_HOSTNAME": {Priority: 0, DisplayName: "Database Hostname", Consumed: true},
+		"DATABASE_NAME":     {Priority: 0, DisplayName: "Database Name", Consumed: true},
+		"DATABASE_USERNAME": {Priority: 0, DisplayName: "Database Username", Consumed: true, Sensitive: true},
+	},
+	Requires: []Stack{
+		DHIS2DB,
+	},
+	Components: []kube.Component{
+		PgAdminComponent{BaseComponent: kube.BaseComponent{Name: "pgadmin"}},
+	},
+}
+
+var pgAdminDefaults = struct {
+	chartVersion string
+}{
+	chartVersion: "1.33.3",
+}
+
+// Stack representing ../../stacks/whoami-go/helmfile.yaml.gotmpl
