+++ pkg/stack/stack.go
@@ -12,8 +12,6 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/dhis2-sre/im-manager/pkg/kube"
-	"github.com/dhis2-sre/im-manager/pkg/model"
 	"github.com/dominikbraun/graph"
 	"golang.org/x/exp/slices"
 	k8s "k8s.io/api/core/v1"
@@ -153,367 +151,6 @@ func ValidateConsumedParameters(stacks []Stack) error {
 const ifNotPresent = "IfNotPresent"
 const always = "Always"
 
-// Stack representing ../../stacks/dhis2-db/helmfile.yaml.gotmpl
-var DHIS2DB = Stack{
-	// TODO: Remove HostnamePattern once stacks 2.0 are the default
-	HostnamePattern: "%s-database-postgresql.%s.svc",
-	Name:            "dhis2-db",
-	Parameters: StackParameters{
-		"DATABASE_ID":               {Priority: 1, DisplayName: "Database"},
-		"DATABASE_SIZE":             {Priority: 2, DisplayName: "Database Size", DefaultValue: &dhis2DBDefaults.dbSize},
-		"DATABASE_NAME":             {Priority: 3, DisplayName: "Database Name", DefaultValue: &dhis2DBDefaults.dbName},
-		"DATABASE_PASSWORD":         {Priority: 4, DisplayName: "Database Password", DefaultValue: &dhis2DBDefaults.dbPassword, Sensitive: true},
-		"DATABASE_USERNAME":         {Priority: 5, DisplayName: "Database Username", DefaultValue: &dhis2DBDefaults.dbUsername, Sensitive: true},
-		"DATABASE_VERSION":          {Priority: 6, DisplayName: "Database Version", DefaultValue: &dhis2DBDefaults.dbVersion},
-		"RESOURCES_REQUESTS_CPU":    {Priority: 7, DisplayName: "Resources Requests CPU", DefaultValue: &dhis2DBDefaults.resourcesRequestsCPU},
-		"RESOURCES_REQUESTS_MEMORY": {Priority: 8, DisplayName: "Resources Requests Memory", DefaultValue: &dhis2DBDefaults.resourcesRequestsMemory},
-		"CHART_VERSION":             {Priority: 9, DisplayName: "Chart Version", DefaultValue: &dhis2DBDefaults.chartVersion},
-	},
-	ParameterProviders: ParameterProviders{
-		"DATABASE_HOSTNAME": postgresHostnameProvider,
-	},
-	Components: []kube.Component{
-		BitnamiPostgresComponent{BaseComponent: kube.BaseComponent{
-			Name:        "db",
-			PVCPatterns: []string{"app.kubernetes.io/instance=%s-database"},
-		}},
-	},
-}
-
-// Provides the PostgreSQL hostname of an instance.
-var postgresHostnameProvider = ParameterProviderFunc(func(instance model.DeploymentInstance) (string, error) {
-	return fmt.Sprintf("%s-%d-database-postgresql.%s.svc", instance.Name, instance.Group.ID, instance.Group.Namespace), nil
-})
-
-var dhis2DBDefaults = struct {
-	chartVersion            string
-	dbID                    string
-	dbName                  string
-	dbPassword              string
-	dbSize                  string
-	dbUsername              string
-	dbVersion               string
-	resourcesRequestsCPU    string
-	resourcesRequestsMemory string
-}{
-	chartVersion:            "16.4.5",
-	dbName:                  "dhis2",
-	dbPassword:              "dhis",
-	dbSize:                  "20Gi",
-	dbUsername:              "dhis",
-	dbVersion:               "16",
-	resourcesRequestsCPU:    "250m",
-	resourcesRequestsMemory: "256Mi",
-}
-
-// Stack representing ../../stacks/minio/helmfile.yaml.gotmpl
-var MINIO = Stack{
-	Name: "minio",
-	Parameters: StackParameters{
-		"MINIO_STORAGE_SIZE":  {Priority: 1, DisplayName: "Storage Size", DefaultValue: &minIODefaults.storageSize},
-		"MINIO_CHART_VERSION": {Priority: 2, DisplayName: "Chart Version", DefaultValue: &minIODefaults.chartVersion},
-		"IMAGE_PULL_POLICY":   {Priority: 3, DisplayName: "Image Pull Policy", DefaultValue: &minIODefaults.imagePullPolicy, Validator: imagePullPolicy},
-		"DATABASE_ID":         {Priority: 0, DisplayName: "Database", Consumed: true},
-	},
-	ParameterProviders: ParameterProviders{
-		"MINIO_HOSTNAME": minioHostnameProvider,
-	},
-	Requires: []Stack{DHIS2DB},
-	Components: []kube.Component{
-		MinioComponent{BaseComponent: kube.BaseComponent{
-			Name:        "minio",
-			PVCPatterns: []string{"app.kubernetes.io/instance=%s-minio"},
-		}},
-	},
-}
-
-// Provides the Minio hostname of an instance.
-var minioHostnameProvider = ParameterProviderFunc(func(instance model.DeploymentInstance) (string, error) {
-	return fmt.Sprintf("%s-minio.%s.svc", instance.Name, instance.Group.Namespace), nil
-})
-
-var storageCompanionProvider = RequireCompanionFunc(func(parameter model.DeploymentInstanceParameter) (*Stack, error) {
-	if parameter.Value == minIOStorage {
-		return &MINIO, nil
-	}
-	return nil, nil
-})
-
-var minIODefaults = struct {
-	chartVersion    string
-	storageSize     string
-	imagePullPolicy string
-}{
-	chartVersion:    "14.7.5",
-	storageSize:     "8Gi",
-	imagePullPolicy: ifNotPresent,
-}
-
-// Stack representing ../../stacks/dhis2-core/helmfile.yaml.gotmpl
-var DHIS2Core = Stack{
-	Name: "dhis2-core",
-	Parameters: StackParameters{
-		"IMAGE_TAG":                       {Priority: 1, DisplayName: "Image Tag", DefaultValue: &dhis2CoreDefaults.imageTag},
-		"IMAGE_REPOSITORY":                {Priority: 2, DisplayName: "Image Repository", DefaultValue: &dhis2CoreDefaults.imageRepository},
-		"IMAGE_PULL_POLICY":               {Priority: 3, DisplayName: "Image Pull Policy", DefaultValue: &dhis2CoreDefaults.imagePullPolicy, Validator: imagePullPolicy},
-		"STORAGE_TYPE":                    {Priority: 4, DisplayName: "Storage type", DefaultValue: &dhis2CoreDefaults.storageType, Validator: storage, RequireCompanion: storageCompanionProvider},
-		"S3_BUCKET":                       {Priority: 5, DisplayName: "S3 bucket", DefaultValue: &dhis2CoreDefaults.s3Bucket},
-		"S3_REGION":                       {Priority: 6, DisplayName: "S3 region", DefaultValue: &dhis2CoreDefaults.s3Region, Sensitive: true},
-		"S3_IDENTITY":                     {Priority: 7, DisplayName: "S3 identity", DefaultValue: &dhis2CoreDefaults.s3Identity, Sensitive: true},
-		"S3_SECRET":                       {Priority: 8, DisplayName: "S3 secret", DefaultValue: &dhis2CoreDefaults.s3Secret, Sensitive: true},
-		"DHIS2_HOME":                      {Priority: 9, DisplayName: "DHIS2 Home Directory", DefaultValue: &dhis2CoreDefaults.dhis2Home},
-		"FLYWAY_MIGRATE_OUT_OF_ORDER":     {Priority: 10, DisplayName: "Flyway Migrate Out Of Order", DefaultValue: &dhis2CoreDefaults.flywayMigrateOutOfOrder},
-		"FLYWAY_REPAIR_BEFORE_MIGRATION":  {Priority: 11, DisplayName: "Flyway Repair Before Migration", DefaultValue: &dhis2CoreDefaults.flywayRepairBeforeMigration},
-		"RESOURCES_REQUESTS_CPU":          {Priority: 12, DisplayName: "Resources Requests CPU", DefaultValue: &dhis2CoreDefaults.resourcesRequestsCPU},
-		"RESOURCES_REQUESTS_MEMORY":       {Priority: 13, DisplayName: "Resources Requests Memory", DefaultValue: &dhis2CoreDefaults.resourcesRequestsMemory},
-		"MIN_READY_SECONDS":               {Priority: 14, DisplayName: "Minimum Ready Seconds", DefaultValue: &dhis2CoreDefaults.minReadySeconds},
-		"LIVENESS_PROBE_TIMEOUT_SECONDS":  {Priority: 15, DisplayName: "Liveness Probe Timeout Seconds", DefaultValue: &dhis2CoreDefaults.livenessProbeTimeoutSeconds},
-		"READINESS_PROBE_TIMEOUT_SECONDS": {Priority: 16, DisplayName: "Readiness Probe Timeout Seconds", DefaultValue: &dhis2CoreDefaults.readinessProbeTimeoutSeconds},
-		"STARTUP_PROBE_FAILURE_THRESHOLD": {Priority: 17, DisplayName: "Startup Probe Failure Threshold", DefaultValue: &dhis2CoreDefaults.startupProbeFailureThreshold},
-		"STARTUP_PROBE_PERIOD_SECONDS":    {Priority: 18, DisplayName: "Startup Probe Period Seconds", DefaultValue: &dhis2CoreDefaults.startupProbePeriodSeconds},
-		"JAVA_OPTS":                       {Priority: 19, DisplayName: "JAVA_OPTS", DefaultValue: &dhis2CoreDefaults.javaOpts},
-		"CHART_VERSION":                   {Priority: 20, DisplayName: "Chart Version", DefaultValue: &dhis2CoreDefaults.chartVersion},
-		"ENABLE_QUERY_LOGGING":            {Priority: 21, DisplayName: "Enable Query Logging", DefaultValue: &dhis2CoreDefaults.enableQueryLogging},
-		"FILESYSTEM_VOLUME_SIZE":          {Priority: 22, DisplayName: "Filesystem volume size (only in effect if \"Storage\" is set to \"filesystem\")", DefaultValue: &dhis2CoreDefaults.filesystemVolumeSize, Sensitive: true},
-		"SAME_SITE_COOKIES":               {Priority: 23, DisplayName: "Same site cookies", DefaultValue: &dhis2CoreDefaults.sameSiteCookies, Validator: sameSiteCookies},
-		"CUSTOM_DHIS2_CONFIG":             {Priority: 24, DisplayName: "Custom DHIS2 config (applied to top of dhis.conf)", DefaultValue: &dhis2CoreDefaults.customDhis2Config, Sensitive: true},
-		"ALLOW_SUSPEND":                   {Priority: 25, DisplayName: "Allow the application to be suspended", DefaultValue: &dhis2CoreDefaults.allowSuspend},
-		"DEPLOY_GLOWROOT":                 {Priority: 26, DisplayName: "Deploy Glowroot", DefaultValue: &dhis2CoreDefaults.deployGlowroot},
-		"DEPLOY_CHAP":                     {Priority: 27, DisplayName: "Deploy CHAP", DefaultValue: &dhis2CoreDefaults.deployChap},
-		"GOOGLE_AUTH_PROJECT_ID":          {Priority: 0, DisplayName: "Google auth project id", DefaultValue: &dhis2CoreDefaults.googleAuthClientId, Sensitive: true},
-		"GOOGLE_AUTH_PRIVATE_KEY":         {Priority: 0, DisplayName: "Google auth private key", DefaultValue: &dhis2CoreDefaults.googleAuthPrivateKey, Sensitive: true},
-		"GOOGLE_AUTH_PRIVATE_KEY_ID":      {Priority: 0, DisplayName: "Google auth private key id", DefaultValue: &dhis2CoreDefaults.googleAuthPrivateKeyId, Sensitive: true},
-		"GOOGLE_AUTH_CLIENT_EMAIL":        {Priority: 0, DisplayName: "Google auth client email", DefaultValue: &dhis2CoreDefaults.googleAuthClientEmail, Sensitive: true},
-		"GOOGLE_AUTH_CLIENT_ID":           {Priority: 0, DisplayName: "Google auth client id", DefaultValue: &dhis2CoreDefaults.googleAuthClientId, Sensitive: true},
-		"DATABASE_HOSTNAME":               {Priority: 0, DisplayName: "Database Hostname", Consumed: true},
-		"DATABASE_NAME":                   {Priority: 0, DisplayName: "Database Name", Consumed: true},
-		"DATABASE_PASSWORD":               {Priority: 0, DisplayName: "Database Password", Consumed: true, Sensitive: true},
-		"DATABASE_USERNAME":               {Priority: 0, DisplayName: "Database Username", Consumed: true, Sensitive: true},
-	},
-	Requires: []Stack{
-		DHIS2DB,
-	},
-	Companions: []Stack{
-		MINIO,
-		ChapCore,
-	},
-	Components: []kube.Component{
-		DHIS2CoreComponent{BaseComponent: kube.BaseComponent{
-			Name: "dhis2",
-			PVCPatterns: []string{
-				"app.kubernetes.io/instance=%s",
-				"app.kubernetes.io/instance=%s-minio",
-			},
-			// All STORAGE_TYPE backends (minio, filesystem, s3) support filestore backup, so no
-			// predicate; the first parameter-gated capability arrives with CNPG-backed stacks.
-			Capabilities: []kube.Capability{
-				{Operation: kube.OperationFilestoreBackup},
-			},
-		}},
-	},
-}
-
-var dhis2CoreDefaults = struct {
-	chartVersion                 string
-	storageType                  string
-	sameSiteCookies              string
-	filesystemVolumeSize         string
-	s3Bucket                     string
-	s3Region                     string
-	s3Identity                   string
-	s3Secret                     string
-	dhis2Home                    string
-	flywayMigrateOutOfOrder      string
-	flywayRepairBeforeMigration  string
-	enableQueryLogging           string
-	imagePullPolicy              string
-	imageRepository              string
-	imageTag                     string
-	javaOpts                     string
-	minReadySeconds              string
-	livenessProbeTimeoutSeconds  string
-	readinessProbeTimeoutSeconds string
-	resourcesRequestsCPU         string
-	resourcesRequestsMemory      string
-	startupProbeFailureThreshold string
-	startupProbePeriodSeconds    string
-	customDhis2Config            string
-	allowSuspend                 string
-	deployGlowroot               string
-	deployChap                   string
-	googleAuthProjectId          string
-	googleAuthPrivateKey         string
-	googleAuthPrivateKeyId       string
-	googleAuthClientEmail        string
-	googleAuthClientId           string
-}{
-	chartVersion:                 "0.34.11",
-	storageType:                  minIOStorage,
-	sameSiteCookies:              lax,
-	filesystemVolumeSize:         "8Gi",
-	s3Bucket:                     "dhis2",
-	s3Region:                     "eu-west-1",
-	s3Identity:                   "-",
-	s3Secret:                     "-",
-	dhis2Home:                    "/opt/dhis2",
-	flywayMigrateOutOfOrder:      "false",
-	flywayRepairBeforeMigration:  "false",
-	enableQueryLogging:           "false",
-	imagePullPolicy:              ifNotPresent,
-	imageRepository:              "core",
-	imageTag:                     "2.40.2",
-	javaOpts:                     " ", // " " is used here since an empty string would be interpreted by helmfile as the environment variable not being set. And since all variables are required an empty string would result in an error
-	minReadySeconds:              "5",
-	livenessProbeTimeoutSeconds:  "1",
-	readinessProbeTimeoutSeconds: "1",
-	resourcesRequestsCPU:         "250m",
-	resourcesRequestsMemory:      "1500Mi",
-	startupProbeFailureThreshold: "26",
-	startupProbePeriodSeconds:    "5",
-	customDhis2Config:            " ",
-	allowSuspend:                 "true",
-	deployGlowroot:               "false",
-	deployChap:                   "false",
-	googleAuthProjectId:          " ", // TODO: " " doesn't need to be used here as with `javaOpts` since the googleAuth* parameters are stack parameters and therefor always populated
-	googleAuthPrivateKey:         " ", // However the web client currently doesn't support these empty parameter so for now
-	googleAuthPrivateKeyId:       " ",
-	googleAuthClientEmail:        " ",
-	googleAuthClientId:           " ",
-}
-
-// Stack representing ../../stacks/dhis2/helmfile.yaml.gotmpl
-var DHIS2 = Stack{
-	// TODO: Remove HostnamePattern once stacks 2.0 are the default
-	HostnamePattern: "%s-database-postgresql.%s.svc",
-	Name:            "dhis2",
-	Parameters: StackParameters{
-		"IMAGE_TAG":                       {Priority: 1, DisplayName: "Image Tag", DefaultValue: &dhis2CoreDefaults.imageTag},
-		"IMAGE_REPOSITORY":                {Priority: 2, DisplayName: "Image Repository", DefaultValue: &dhis2CoreDefaults.imageRepository},
-		"IMAGE_PULL_POLICY":               {Priority: 3, DisplayName: "Image Pull Policy", DefaultValue: &dhis2CoreDefaults.imagePullPolicy, Validator: imagePullPolicy},
-		"DATABASE_ID":                     {Priority: 4, DisplayName: "Database"},
-		"DATABASE_NAME":                   {Priority: 5, DisplayName: "Database Name", DefaultValue: &dhis2DBDefaults.dbName},
-		"DATABASE_PASSWORD":               {Priority: 6, DisplayName: "Database Password", DefaultValue: &dhis2DBDefaults.dbPassword, Sensitive: true},
-		"DATABASE_SIZE":                   {Priority: 7, DisplayName: "Database Size", DefaultValue: &dhis2DBDefaults.dbSize},
-		"DATABASE_USERNAME":               {Priority: 8, DisplayName: "Database Username", DefaultValue: &dhis2DBDefaults.dbUsername, Sensitive: true},
-		"DATABASE_VERSION":                {Priority: 9, DisplayName: "Database Version", DefaultValue: &dhis2DBDefaults.dbVersion},
-		"INSTALL_REDIS":                   {Priority: 10, DisplayName: "Install Redis", DefaultValue: &dhis2Defaults.installRedis},
-		"DHIS2_HOME":                      {Priority: 11, DisplayName: "DHIS2 Home Directory", DefaultValue: &dhis2CoreDefaults.dhis2Home},
-		"FLYWAY_MIGRATE_OUT_OF_ORDER":     {Priority: 12, DisplayName: "Flyway Migrate Out Of Order", DefaultValue: &dhis2CoreDefaults.flywayMigrateOutOfOrder},
-		"FLYWAY_REPAIR_BEFORE_MIGRATION":  {Priority: 13, DisplayName: "Flyway Repair Before Migration", DefaultValue: &dhis2CoreDefaults.flywayRepairBeforeMigration},
-		"CORE_RESOURCES_REQUESTS_CPU":     {Priority: 14, DisplayName: "Core Resources Requests CPU", DefaultValue: &dhis2CoreDefaults.resourcesRequestsCPU},
-		"CORE_RESOURCES_REQUESTS_MEMORY":  {Priority: 15, DisplayName: "Core Resources Requests Memory", DefaultValue: &dhis2CoreDefaults.resourcesRequestsMemory},
-		"DB_RESOURCES_REQUESTS_CPU":       {Priority: 16, DisplayName: "DB Resources Requests CPU", DefaultValue: &dhis2DBDefaults.resourcesRequestsCPU},
-		"DB_RESOURCES_REQUESTS_MEMORY":    {Priority: 17, DisplayName: "DB Resources Requests Memory", DefaultValue: &dhis2DBDefaults.resourcesRequestsMemory},
-		"MIN_READY_SECONDS":               {Priority: 18, DisplayName: "Minimum Ready Seconds", DefaultValue: &dhis2CoreDefaults.minReadySeconds},
-		"LIVENESS_PROBE_TIMEOUT_SECONDS":  {Priority: 19, DisplayName: "Liveness Probe Timeout Seconds", DefaultValue: &dhis2CoreDefaults.livenessProbeTimeoutSeconds},
-		"READINESS_PROBE_TIMEOUT_SECONDS": {Priority: 20, DisplayName: "Readiness Probe Timeout Seconds", DefaultValue: &dhis2CoreDefaults.readinessProbeTimeoutSeconds},
-		"STARTUP_PROBE_FAILURE_THRESHOLD": {Priority: 21, DisplayName: "Startup Probe Failure Threshold", DefaultValue: &dhis2CoreDefaults.startupProbeFailureThreshold},
-		"STARTUP_PROBE_PERIOD_SECONDS":    {Priority: 22, DisplayName: "Startup Probe Period Seconds", DefaultValue: &dhis2CoreDefaults.startupProbePeriodSeconds},
-		"CHART_VERSION":                   {Priority: 23, DisplayName: "Chart Version", DefaultValue: &dhis2CoreDefaults.chartVersion},
-		"JAVA_OPTS":                       {Priority: 24, DisplayName: "JAVA Options", DefaultValue: &dhis2CoreDefaults.javaOpts},
-		"ENABLE_QUERY_LOGGING":            {Priority: 25, DisplayName: "Enable Query Logging", DefaultValue: &dhis2CoreDefaults.enableQueryLogging},
-		"GOOGLE_AUTH_PROJECT_ID":          {Priority: 0, DisplayName: "Google auth project id", DefaultValue: &dhis2CoreDefaults.googleAuthClientId, Sensitive: true},
-		"GOOGLE_AUTH_PRIVATE_KEY":         {Priority: 0, DisplayName: "Google auth private key", DefaultValue: &dhis2CoreDefaults.googleAuthPrivateKey, Sensitive: true},
-		"GOOGLE_AUTH_PRIVATE_KEY_ID":      {Priority: 0, DisplayName: "Google auth private key id", DefaultValue: &dhis2CoreDefaults.googleAuthPrivateKeyId, Sensitive: true},
-		"GOOGLE_AUTH_CLIENT_EMAIL":        {Priority: 0, DisplayName: "Google auth client email", DefaultValue: &dhis2CoreDefaults.googleAuthClientEmail, Sensitive: true},
-		"GOOGLE_AUTH_CLIENT_ID":           {Priority: 0, DisplayName: "Google auth client id", DefaultValue: &dhis2CoreDefaults.googleAuthClientId, Sensitive: true},
-	},
-	ParameterProviders: ParameterProviders{
-		"DATABASE_HOSTNAME": postgresHostnameProvider,
-	},
-	Components: []kube.Component{
-		DHIS2CoreComponent{BaseComponent: kube.BaseComponent{Name: "dhis2"}},
-		BitnamiPostgresComponent{BaseComponent: kube.BaseComponent{
-			Name: "db",
-			PVCPatterns: []string{
-				"app.kubernetes.io/instance=%s-database",
-				// The redis release carries no im labels yet, so it has no component; its PVC
-				// still rides here until redis labels land (roadmap step 4).
-				"app.kubernetes.io/instance=%s-redis",
-			},
-		}},
-	},
-}
-
-// DHIS2V2 deploys the dhis2 umbrella chart: one release bundling DHIS 2 core, PostgreSQL and an
-// optional MinIO file store, replacing the deployment-level composition of dhis2-db, dhis2-core
-// and the minio companion.
-var DHIS2V2 = withGroupedParameters(Stack{
-	Name: "dhis2-v2",
-	ParameterGroups: []ParameterGroup{
-		{Name: "dhis2", Title: "DHIS 2 Core", Parameters: StackParameters{
-			"IMAGE_TAG":                       {Priority: 1, DisplayName: "Image Tag", DefaultValue: &dhis2CoreDefaults.imageTag},
-			"IMAGE_REPOSITORY":                {Priority: 2, DisplayName: "Image Repository", DefaultValue: &dhis2CoreDefaults.imageRepository},
-			"IMAGE_PULL_POLICY":               {Priority: 3, DisplayName: "Image Pull Policy", DefaultValue: &dhis2CoreDefaults.imagePullPolicy, Validator: imagePullPolicy},
-			"DHIS2_HOME":                      {Priority: 17, DisplayName: "DHIS2 Home Directory", DefaultValue: &dhis2CoreDefaults.dhis2Home},
-			"FLYWAY_MIGRATE_OUT_OF_ORDER":     {Priority: 18, DisplayName: "Flyway Migrate Out Of Order", DefaultValue: &dhis2CoreDefaults.flywayMigrateOutOfOrder},
-			"FLYWAY_REPAIR_BEFORE_MIGRATION":  {Priority: 19, DisplayName: "Flyway Repair Before Migration", DefaultValue: &dhis2CoreDefaults.flywayRepairBeforeMigration},
-			"CORE_RESOURCES_REQUESTS_CPU":     {Priority: 20, DisplayName: "Resources Requests CPU", DefaultValue: &dhis2CoreDefaults.resourcesRequestsCPU},
-			"CORE_RESOURCES_REQUESTS_MEMORY":  {Priority: 21, DisplayName: "Resources Requests Memory", DefaultValue: &dhis2CoreDefaults.resourcesRequestsMemory},
-			"MIN_READY_SECONDS":               {Priority: 24, DisplayName: "Minimum Ready Seconds", DefaultValue: &dhis2CoreDefaults.minReadySeconds},
-			"LIVENESS_PROBE_TIMEOUT_SECONDS":  {Priority: 25, DisplayName: "Liveness Probe Timeout Seconds", DefaultValue: &dhis2CoreDefaults.livenessProbeTimeoutSeconds},
-			"READINESS_PROBE_TIMEOUT_SECONDS": {Priority: 26, DisplayName: "Readiness Probe Timeout Seconds", DefaultValue: &dhis2CoreDefaults.readinessProbeTimeoutSeconds},
-			"STARTUP_PROBE_FAILURE_THRESHOLD": {Priority: 27, DisplayName: "Startup Probe Failure Threshold", DefaultValue: &dhis2CoreDefaults.startupProbeFailureThreshold},
-			"STARTUP_PROBE_PERIOD_SECONDS":    {Priority: 28, DisplayName: "Startup Probe Period Seconds", DefaultValue: &dhis2CoreDefaults.startupProbePeriodSeconds},
-			"CHART_VERSION":                   {Priority: 29, DisplayName: "Chart Version", DefaultValue: &dhis2V2Defaults.chartVersion},
-			"JAVA_OPTS":                       {Priority: 30, DisplayName: "JAVA Options", DefaultValue: &dhis2CoreDefaults.javaOpts},
-			"ENABLE_QUERY_LOGGING":            {Priority: 31, DisplayName: "Enable Query Logging", DefaultValue: &dhis2CoreDefaults.enableQueryLogging},
-			"SAME_SITE_COOKIES":               {Priority: 32, DisplayName: "Same site cookies", DefaultValue: &dhis2CoreDefaults.sameSiteCookies, Validator: sameSiteCookies},
-			"CUSTOM_DHIS2_CONFIG":             {Priority: 33, DisplayName: "Custom DHIS2 config (applied to top of dhis.conf)", DefaultValue: &dhis2CoreDefaults.customDhis2Config, Sensitive: true},
-			"GOOGLE_AUTH_PROJECT_ID":          {Priority: 34, DisplayName: "Google auth project id", DefaultValue: &dhis2CoreDefaults.googleAuthProjectId, Sensitive: true},
-			"GOOGLE_AUTH_PRIVATE_KEY":         {Priority: 35, DisplayName: "Google auth private key", DefaultValue: &dhis2CoreDefaults.googleAuthPrivateKey, Sensitive: true},
-			"GOOGLE_AUTH_PRIVATE_KEY_ID":      {Priority: 36, DisplayName: "Google auth private key id", DefaultValue: &dhis2CoreDefaults.googleAuthPrivateKeyId, Sensitive: true},
-			"GOOGLE_AUTH_CLIENT_EMAIL":        {Priority: 37, DisplayName: "Google auth client email", DefaultValue: &dhis2CoreDefaults.googleAuthClientEmail, Sensitive: true},
-			"GOOGLE_AUTH_CLIENT_ID":           {Priority: 38, DisplayName: "Google auth client id", DefaultValue: &dhis2CoreDefaults.googleAuthClientId, Sensitive: true},
-		}},
-		{Name: "db", Title: "PostgreSQL", Parameters: StackParameters{
-			"DATABASE_ID":                  {Priority: 4, DisplayName: "Database"},
-			"DATABASE_NAME":                {Priority: 5, DisplayName: "Database Name", DefaultValue: &dhis2DBDefaults.dbName},
-			"DATABASE_PASSWORD":            {Priority: 6, DisplayName: "Database Password", DefaultValue: &dhis2DBDefaults.dbPassword, Sensitive: true},
-			"DATABASE_SIZE":                {Priority: 7, DisplayName: "Database Size", DefaultValue: &dhis2DBDefaults.dbSize},
-			"DATABASE_USERNAME":            {Priority: 8, DisplayName: "Database Username", DefaultValue: &dhis2DBDefaults.dbUsername, Sensitive: true},
-			"DB_RESOURCES_REQUESTS_CPU":    {Priority: 22, DisplayName: "Resources Requests CPU", DefaultValue: &dhis2DBDefaults.resourcesRequestsCPU},
-			"DB_RESOURCES_REQUESTS_MEMORY": {Priority: 23, DisplayName: "Resources Requests Memory", DefaultValue: &dhis2DBDefaults.resourcesRequestsMemory},
-		}},
-		{Name: "storage", Title: "Storage", Parameters: StackParameters{
-			"STORAGE_TYPE": {Priority: 10, DisplayName: "Storage type", DefaultValue: &dhis2CoreDefaults.storageType, Validator: storage},
-		}},
-		{Name: "minio", Title: "MinIO", When: whenStorageIsMinio, Parameters: StackParameters{
-			"MINIO_STORAGE_SIZE": {Priority: 11, DisplayName: "Storage size", DefaultValue: &minIODefaults.storageSize},
-		}},
-		{Name: "s3", Title: "S3", When: whenStorageIsS3, Parameters: StackParameters{
-			"S3_BUCKET":   {Priority: 13, DisplayName: "Bucket", DefaultValue: &dhis2CoreDefaults.s3Bucket},
-			"S3_REGION":   {Priority: 14, DisplayName: "Region", DefaultValue: &dhis2CoreDefaults.s3Region, Sensitive: true},
-			"S3_IDENTITY": {Priority: 15, DisplayName: "Identity", DefaultValue: &dhis2CoreDefaults.s3Identity, Sensitive: true},
-			"S3_SECRET":   {Priority: 16, DisplayName: "Secret", DefaultValue: &dhis2CoreDefaults.s3Secret, Sensitive: true},
-		}},
-		{Name: "filesystem", Title: "Filesystem", When: whenStorageIsFilesystem, Parameters: StackParameters{
-			"FILESYSTEM_VOLUME_SIZE": {Priority: 12, DisplayName: "Volume size", DefaultValue: &dhis2CoreDefaults.filesystemVolumeSize, Sensitive: true},
-		}},
-	},
-	ParameterProviders: ParameterProviders{
-		"DATABASE_HOSTNAME": dhis2V2PostgresHostnameProvider,
-	},
-	Components: []kube.Component{
-		DHIS2CoreComponent{BaseComponent: kube.BaseComponent{
-			Name:        "dhis2",
-			PVCPatterns: []string{"app.kubernetes.io/instance=%s,app.kubernetes.io/name=dhis2"},
-			Capabilities: []kube.Capability{
-				{Operation: kube.OperationFilestoreBackup},
-			},
-		}},
-		CNPGPostgresComponent{BaseComponent: kube.BaseComponent{
-			Name:        "db",
-			PVCPatterns: []string{"cnpg.io/cluster=%s-dhis2-postgresql"},
-		}, ClusterPattern: "%s-dhis2-postgresql"},
-		MinioComponent{BaseComponent: kube.BaseComponent{
-			Name:        "minio",
-			PVCPatterns: []string{"app.kubernetes.io/instance=%s,app.kubernetes.io/name=minio"},
-			When:        whenStorageIsMinio,
-		}},
-	},
-})
-
 // withGroupedParameters flattens the parameters declared inside a stack's groups into the stack's
 // parameter map, stamping each parameter's Group with the declaring group's name, so membership
 // follows from where a parameter is declared instead of a string reference. Panics on a duplicate
@@ -534,120 +171,6 @@ func withGroupedParameters(s Stack) Stack {
 	return s
 }
 
-var dhis2V2Defaults = struct {
-	chartVersion string
-}{
-	chartVersion: "1.0.0",
-}
-
-// The storage conditions mirror the chart's own gating of each file store backend on STORAGE_TYPE.
-// The minio condition doubles as the minio component's presence predicate, so the deploy form's
-// section visibility and the backend's component listing cannot disagree.
-var (
-	whenStorageIsMinio      = &kube.Condition{Parameter: "STORAGE_TYPE", Equals: minIOStorage}
-	whenStorageIsS3         = &kube.Condition{Parameter: "STORAGE_TYPE", Equals: s3Storage}
-	whenStorageIsFilesystem = &kube.Condition{Parameter: "STORAGE_TYPE", Equals: filesystemStorage}
-)
-
-// Provides the PostgreSQL hostname of a dhis2-v2 instance: the CNPG cluster's read-write service,
-// named after the chart fullname (<release>-dhis2).
-var dhis2V2PostgresHostnameProvider = ParameterProviderFunc(func(instance model.DeploymentInstance) (string, error) {
-	return fmt.Sprintf("%s-%d-dhis2-postgresql-rw.%s.svc", instance.Name, instance.Group.ID, instance.Group.Namespace), nil
-})
-
-var dhis2Defaults = struct {
-	installRedis string
-}{
-	installRedis: "false",
-}
-
-// Stack representing ../../stacks/pgadmin/helmfile.yaml.gotmpl
-var PgAdmin = Stack{
-	Name: "pgadmin",
-	Parameters: StackParameters{
-		"PGADMIN_USERNAME":  {Priority: 1, DisplayName: "pgAdmin Username", Sensitive: true},
-		"PGADMIN_PASSWORD":  {Priority: 2, DisplayName: "pgAdmin Password", Sensitive: true},
-		"CHART_VERSION":     {Priority: 3, DisplayName: "Chart Version", DefaultValue: &pgAdminDefaults.chartVersion},
-		"DATABASE_HOSTNAME": {Priority: 0, DisplayName: "Database Hostname", Consumed: true},
-		"DATABASE_NAME":     {Priority: 0, DisplayName: "Database Name", Consumed: true},
-		"DATABASE_USERNAME": {Priority: 0, DisplayName: "Database Username", Consumed: true, Sensitive: true},
-	},
-	Requires: []Stack{
-		DHIS2DB,
-	},
-	Components: []kube.Component{
-		PgAdminComponent{BaseComponent: kube.BaseComponent{Name: "pgadmin"}},
-	},
-}
-
-var pgAdminDefaults = struct {
-	chartVersion string
-}{
-	chartVersion: "1.33.3",
-}
-
-// Stack representing ../../stacks/whoami-go/helmfile.yaml.gotmpl
-var WhoamiGo = Stack{
-	Name: "whoami-go",
-	Parameters: StackParameters{
-		"IMAGE_TAG":         {Priority: 1, DisplayName: "Image Tag", DefaultValue: &whoamiGoDefaults.imageTag},
-		"IMAGE_REPOSITORY":  {Priority: 2, DisplayName: "Image Repository", DefaultValue: &whoamiGoDefaults.imageRepository},
-		"IMAGE_PULL_POLICY": {Priority: 3, DisplayName: "Image Pull Policy", DefaultValue: &whoamiGoDefaults.imagePullPolicy, Validator: imagePullPolicy},
-		"REPLICA_COUNT":     {Priority: 4, DisplayName: "Replica Count", DefaultValue: &whoamiGoDefaults.replicaCount},
-		"CHART_VERSION":     {Priority: 5, DisplayName: "Chart Version", DefaultValue: &whoamiGoDefaults.chartVersion},
-	},
-	Components: []kube.Component{
-		WhoamiComponent{BaseComponent: kube.BaseComponent{Name: "whoami"}},
-	},
-}
-
-var whoamiGoDefaults = struct {
-	chartVersion    string
-	imagePullPolicy string
-	imageRepository string
-	imageTag        string
-	replicaCount    string
-}{
-	chartVersion:    "0.9.0",
-	imagePullPolicy: ifNotPresent,
-	imageRepository: "whoami-go",
-	imageTag:        "0.6.0",
-	replicaCount:    "1",
-}
-
-// Stack representing ../../stacks/im-job-runner/helmfile.yaml.gotmpl
-var IMJobRunner = Stack{
-	Name: "im-job-runner",
-	Parameters: StackParameters{
-		"COMMAND":                 {Priority: 0, DisplayName: "Command"},
-		"PAYLOAD":                 {Priority: 0, DisplayName: "Payload", DefaultValue: &imJobRunnerDefaults.payload},
-		"DHIS2_DATABASE_DATABASE": {Priority: 0, DisplayName: "DHIS2 Database Name", DefaultValue: &dhis2DBDefaults.dbName},
-		"DHIS2_DATABASE_HOSTNAME": {Priority: 0, DisplayName: "DHIS2 Database Hostname", DefaultValue: &imJobRunnerDefaults.dbHostname},
-		"DHIS2_DATABASE_PASSWORD": {Priority: 0, DisplayName: "DHIS2 Database Password", DefaultValue: &dhis2DBDefaults.dbPassword, Sensitive: true},
-		"DHIS2_DATABASE_PORT":     {Priority: 0, DisplayName: "DHIS2 Database Port", DefaultValue: &imJobRunnerDefaults.dbPort},
-		"DHIS2_DATABASE_USERNAME": {Priority: 0, DisplayName: "DHIS2 Database Username", DefaultValue: &dhis2DBDefaults.dbUsername, Sensitive: true},
-		"DHIS2_HOSTNAME":          {Priority: 0, DisplayName: "DHIS2 Hostname", DefaultValue: &imJobRunnerDefaults.dhis2Hostname},
-		"CHART_VERSION":           {Priority: 0, DisplayName: "Chart Version", DefaultValue: &imJobRunnerDefaults.chartVersion},
-	},
-	// No components: the job-runner is an old jobs experiment that only labels pods, so no
-	// workload is addressable. Jobs get their own design in a separate task.
-}
-
-var imJobRunnerDefaults = struct {
-	chartVersion  string
-	dbHostname    string
-	dbPort        string
-	dhis2Hostname string
-	payload       string
-}{
-	chartVersion:  "0.1.0",
-	dbHostname:    "-",
-	dbPort:        "5432",
-	dhis2Hostname: "-",
-	payload:       "-",
-}
-
-// imagePullPolicy validates a value is a valid Kubernetes image pull policy.
 var imagePullPolicy = OneOf(string(k8s.PullAlways), string(k8s.PullNever), string(k8s.PullIfNotPresent))
 
 const (
